changed the scale of the screen
Things were working on a screen of 450x600 before and the snake was moving a pace of 10 px per cycle. I changed that to one move per cycle and change the size of each node to 1x1. I then multiplied all the size and positions by a scaler value which maps it to the canvas. This allows greater flexibility into the feature.